Made uploads filename more reasonably.
Uploaded images are now stored in
STATIC_ROOT/uploads/YYYY/uploadName.extension
instead of STATIC_ROOT/uploads/imageFileName.extension, where YYYY are
the year the images were uploaded.

This is only the case when there is no other file in that directory with
the same name and extension. In that case there will be some random
characters appended to the filename, which more or less destroys the
purpose of this commit.

Maybe we will get to implement the route functionality soon...