Layout and specifically titles are more DRY.
Before we rewrote basically the whole title for every page, even though
all were of the format PAGENAME - DNollK YEAR, where YEAR was hardcoded
string like 2015.

This is error prone as we would only need to miss one page to make an
inconsistent and rather poor impression on the viewer.
Now, however, in the base layout I've defined the title as PAGE - DNollK
YEAR, where YEAR comes from a template tag returning the current year.
PAGE is a new block which means it can be shadowed by each template.