Improved animation for sidebar
Now the navigation sidebar will be hidden from 0 - 549px instead of beginning to be visible at 450px. Same with the content.

The navigation timings are improved as well by using delay instead of animating for 1s on showing and 0.7s on hiding. Now the animation is always 0.7s but has a 0.1s delay when showing.