Implemented pagination for news.
This commit is related to issue #14.

Apparently Django has inbuilt support for pagination, so the
implementation was much easier than I anticipated.
I would have liked to only retrieve the results that I was interested in
from the DB, but this was the official and documented way to do it.